library/requirements.txt
flake8==3.8.4
numpy==1.19.5
Pillow==8.1.0
wheel=0.36.2
setuptools=52.0.0
twine=3.3.0
